Refactor to individual converters
