+++ shared/i18n/locales/en_US/translation.json
@@ -75,6 +75,7 @@
   "Download document": "Download document",
   "Download as Markdown": "Download as Markdown",
   "Download as HTML": "Download as HTML",
+  "Download as TextBundle": "Download as TextBundle",
   "Download as PDF": "Download as PDF",
   "Copy as Markdown": "Copy as Markdown",
   "Markdown copied to clipboard": "Markdown copied to clipboard",
@@ -280,6 +281,7 @@
   "Preparing your download": "Preparing your download",
   "A file containing the selected documents in Markdown format.": "A file containing the selected documents in Markdown format.",
   "A file containing the selected documents in HTML format.": "A file containing the selected documents in HTML format.",
+  "A file containing the selected documents and their images in TextBundle format.": "A file containing the selected documents and their images in TextBundle format.",
   "A file containing the selected documents in PDF format.": "A file containing the selected documents in PDF format.",
   "Include child documents": "Include child documents",
   "When selected, exporting the document <em>{{documentName}}</em> may take some time.": "When selected, exporting the document <em>{{documentName}}</em> may take some time.",
@@ -358,6 +360,7 @@
   "Show detail": "Show detail",
   "A ZIP file containing the images, and documents in the Markdown format.": "A ZIP file containing the images, and documents in the Markdown format.",
   "A ZIP file containing the images, and documents as HTML files.": "A ZIP file containing the images, and documents as HTML files.",
+  "A ZIP file containing each document and its images as a TextBundle.": "A ZIP file containing each document and its images as a TextBundle.",
   "Structured data that can be used to transfer data to another compatible {{ appName }} instance.": "Structured data that can be used to transfer data to another compatible {{ appName }} instance.",
   "Exporting the collection <em>{{collectionName}}</em> may take some time.": "Exporting the collection <em>{{collectionName}}</em> may take some time.",
   "Include attachments": "Include attachments",
